requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
httpx
pypdf
python-docx
python-multipart
